Update submodule version
This updates the submodule to the latest upstream version,
mainly to allow using these bindings on Raspberry Pi models
that were released in the meantime. In particular, I tested
this on a Raspberry Zero W.

This requires some code changes, since the layout of some of
the C structs in the upstream library has changed.

Additionally, I globally replaced the 'libc::uint*_t' types
by rust's built-in 'u*', to silence a few dozen compiler warnings.